Move shared/services/apollo to data/services/apollo and fix paths
