+++ assets/src/data/services/apollo/queries/currentUser/test/data.ts
@@ -1,7 +1,7 @@
 import { GraphQLError } from 'graphql';
 
 import { ReadQueryOptions } from '@dataServices/apollo/queries';
-import { Viewer } from '../../../../../../../application/valueObjects/config/types';
+import { Viewer } from '@application/valueObjects/config/types';
 import { GET_CURRENT_USER } from '..';
 
 export const request: ReadQueryOptions = {
