+++ assets/src/data/services/apollo/queries/currentUser/useCurrentUserQueryOptions.ts
