+++ assets/src/data/services/apollo/queries/generalSettings/index.ts
