+++ assets/src/data/services/apollo/queries/generalSettings/queries.ts
