+++ assets/src/data/services/apollo/queries/generalSettings/test/useFetchGeneralSettings.test.ts
@@ -1,7 +1,7 @@
 import { renderHook } from '@testing-library/react-hooks';
 
 import useFetchGeneralSettings from '../useFetchGeneralSettings';
-import { ApolloMockedProvider } from '../../../../../../eventEditor/services/context/TestContext';
+import { ApolloMockedProvider } from '@edtrServices/context/TestContext';
 import { successMocks, errorMocks, generalSettings } from './data';
 
 const timeout = 5000; // milliseconds
