+++ assets/src/data/services/apollo/queries/generalSettings/useFetchGeneralSettings.ts
@@ -1,7 +1,7 @@
 import { useQuery } from '@apollo/react-hooks';
-import { GET_GENERAL_SETTINGS } from './';
-import { FetchEntitiesResult } from '../../../../../eventEditor/services/apollo/queries/types';
-import { GeneralSettingsData } from '../../../../../../application/valueObjects/config/types';
+import { GET_GENERAL_SETTINGS } from '.';
+import { FetchEntitiesResult } from '@edtrServices/apollo/queries/types';
+import { GeneralSettingsData } from '@application/valueObjects/config/types';
 
 const useFetchGeneralSettings = (): FetchEntitiesResult<GeneralSettingsData> => {
 	const { data, error, loading } = useQuery<GeneralSettingsData>(GET_GENERAL_SETTINGS);
