+++ assets/src/data/services/apollo/queries/generalSettings/useGeneralSettingsQueryOptions.ts
