+++ assets/src/data/services/apollo/queries/generalSettings/useUpdateGeneralSettingsCache.ts
@@ -1,7 +1,6 @@
-import { WriteQueryOptions } from '../../../../../eventEditor/services/apollo/queries/types';
-import { CacheUpdaterFn } from '../types';
+import { CacheUpdaterFn, WriteQueryOptions } from '../types';
 import useUpdateCache from '../useUpdateCache';
-import { GeneralSettingsData } from '../../../../../../application/valueObjects/config/types';
+import { GeneralSettingsData } from '@application/valueObjects/config/types';
 
 const useUpdateGeneralSettingsCache = (
 	writeQueryOptions: WriteQueryOptions<GeneralSettingsData> = undefined
