+++ assets/src/data/services/apollo/queries/index.ts
@@ -1,10 +1,16 @@
 export { default as useCacheQuery } from './useCacheQuery';
 
+export { default as useUpdateCache } from './useUpdateCache';
+
 export * from './attendees';
+
+export * from './currentUser';
 
 export * from './datetimes';
 
 export * from './events';
+
+export * from './generalSettings';
 
 export * from './tickets';
