+++ assets/src/data/services/apollo/queries/types.ts
@@ -1,6 +1,7 @@
 import { QueryHookOptions } from '@apollo/react-hooks';
 import { DocumentNode } from 'graphql';
 import { OperationVariables } from 'apollo-client';
+import { DataProxy } from 'apollo-cache';
 
 export interface EntityQueryArgs<WhereArgs> {
 	after?: string;
@@ -23,3 +24,16 @@ type EntityQueryOrderByItem<Field> = {
 };
 
 export type EntityQueryOrderBy<Field> = Array<EntityQueryOrderByItem<Field>>;
+
+export interface FetchEntitiesResult<Data> {
+	data: Data;
+	error: Error;
+	loading: boolean;
+}
+
+export type CacheUpdaterFn<TData = any> = (writeOptions?: WriteQueryOptions<TData>) => void;
+
+export type WriteQueryOptions<TData = any, TVariables = OperationVariables> = DataProxy.WriteQueryOptions<
+	TData,
+	TVariables
+>;
