+++ assets/src/domain/eventEditor/services/context/TestContext/data.ts
@@ -1,14 +1,14 @@
-import { EventData } from '../../../interfaces/types';
-import { RelationalData } from '../../../../../application/services/apollo/relations';
+import { EventData } from '@edtrInterfaces/types';
+import { RelationalData } from '@appServices/apollo/relations';
 import { nodes as datetimes, edge as datetimesEdge } from '../../apollo/queries/datetimes/test/data';
 
 import { nodes as tickets, edge as ticketsEdge } from '../../apollo/queries/tickets/test/data';
 import { nodes as prices, edge as pricesEdge } from '../../apollo/queries/prices/test/data';
 import { nodes as priceTypes, edge as priceTypesEdge } from '../../apollo/queries/priceTypes/test/data';
 
-export { currentUser } from '../../../../shared/services/apollo/queries/currentUser/test/data';
+export { currentUser } from '@dataServices/apollo/queries/currentUser/test/data';
 
-export { generalSettings } from '../../../../shared/services/apollo/queries/generalSettings/test/data';
+export { generalSettings } from '@dataServices/apollo/queries/generalSettings/test/data';
 
 export const eventId = 100;
