--- assets/src/domain/shared/services/apollo/queries/index.ts
@@ -1,7 +1,0 @@
-export { default as useUpdateCache } from './useUpdateCache';
-
-export * from './currentUser';
-
-export * from './generalSettings';
-
-export * from './types';
