--- assets/src/domain/shared/services/apollo/queries/types.ts
@@ -1,3 +1,0 @@
-import { WriteQueryOptions } from '@edtrServices/apollo/queries/types';
-
-export type CacheUpdaterFn<TData = any> = (writeOptions?: WriteQueryOptions<TData>) => void;
